refactor: move LocationNameProvider to util package
- Changed the package of LocationNameProvider from map to util.
- Updated import statements in related files.
- Made constructor private for singleton pattern.
